Fixing a warning

+++ Marvel API/Networking Extensions/MDAMarvelAPIClient.m
@@ -31,7 +31,7 @@ static NSString * const CBPMarvelAPIBaseURLString = @"http://gateway.marvel.com"
     const char *ptr = [[NSString stringWithFormat:@"%@%@%@", timestamp, CBPMarvelAPIPrivateKey, CBPMarvelAPIPublicKey] UTF8String];
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
     
-    CC_MD5(ptr, strlen(ptr), md5Buffer);
+    CC_MD5(ptr, (int)strlen(ptr), md5Buffer);
     
     NSMutableString *hash = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(int i = 0; i < CC_MD5_DIGEST_LENGTH; i++) {
